Remove errors when building cld2 with gcc -std=c++11
Apparently, narrowing conversions are disallowed in C++11. This line
turns the errors that pop up when compiling cld2 with a C++11-adhering
GCC into warnings.

I dedicate any and all copyright interest in my contributions to this
project to the public domain. I make this dedication for the benefit of
the public at large and to the detriment of my heirs and successors. I
intend this dedication to be an overt act of relinquishment in
perpetuity of all present and future rights to this software under
copyright law.